cleanup common and improve error handling some
