+++ node/coinstacks/common/api/src/evm/index.ts
@@ -1,3 +1,3 @@
 export * from './models'
-export * from './utils'
 export * from './gasOracle'
+export { formatAddress } from './service'
