+++ node/packages/blockbook/src/models.ts
@@ -5,10 +5,9 @@ import axios, { AxiosError } from 'axios'
  * Generic api error for handling failed requests
  */
 export class ApiError extends AxiosError {
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  constructor(err: AxiosError<any> | Error) {
+  constructor(err: AxiosError | Error) {
     if (axios.isAxiosError(err)) {
-      super(err.response?.data?.error ?? JSON.stringify(err.response?.data) ?? err.message)
+      super(err.response?.data?.error || JSON.stringify(err.response?.data) || err.message)
     } else {
       super(err.message)
     }
